refactor: :truck: TaxJarTaxCodeMatcher -> TaxJarCalculateTaxesTaxCodeMatcher
