+++ apps/taxes/src/modules/taxjar/calculate-taxes/taxjar-calculate-taxes-payload-lines-transformer.ts
@@ -4,7 +4,7 @@ import {
   TaxJarCalculateTaxesPayload,
   TaxJarCalculateTaxesTarget,
 } from "./taxjar-calculate-taxes-adapter";
-import { TaxJarTaxCodeMatcher } from "../tax-code/taxjar-tax-code-matcher";
+import { TaxJarCalculateTaxesTaxCodeMatcher } from "./taxjar-calculate-taxes-tax-code-matcher";
 
 export class TaxJarCalculateTaxesPayloadLinesTransformer {
   transform(
@@ -21,7 +21,7 @@ export class TaxJarCalculateTaxesPayloadLinesTransformer {
 
     const mappedLines: TaxJarCalculateTaxesTarget["params"]["line_items"] = lines.map(
       (line, index) => {
-        const matcher = new TaxJarTaxCodeMatcher();
+        const matcher = new TaxJarCalculateTaxesTaxCodeMatcher();
         const discountAmount = distributedDiscounts[index];
         const taxCode = matcher.match(line, matches);
